+++ wparc/cmds/extractor.py
@@ -2,7 +2,7 @@
 
 import logging
 
-from ..wpapi.crawler import collect_data, collect_files
+from ..wpapi.crawler import collect_data, collect_files, ping
 
 
 class Project:
@@ -12,11 +12,14 @@ class Project:
         pass
 
 
-    def dump(self, domain, all):
-        collect_data(domain, get_unknown=all)
+    def dump(self, domain, all, https):
+        collect_data(domain, get_unknown=all, force_https=https)
         pass
 
     def getfiles(self, domain):
         collect_files(domain)
         pass
+
+    def ping(self, domain, https):
+        ping(domain, force_https=https)