+++ HISTORY.rst
@@ -4,6 +4,13 @@ History
 =======
 
 
+
+1.0.2 (2022-04-02)
+------------------
+* Added command "ping" to verify existence of /wp-json/ endpoint
+* Added option "https" for "ping" and "dump" commands. It forces using https by default insted of http
+
+
 1.0.1 (2022-03-31)
 ------------------
 * First public release on PyPI and updated github code